activities_agent/requirements.txt
a2a-sdk
python-dotenv
uvicorn
google-adk
nest-asyncio